+++ client/Packages/com.beamable/Editor/UI/Common/Components/UpdateAvailableAnnouncementVisualElement/UpdateAvailableAnnouncementVisualElement.cs
@@ -48,7 +48,11 @@ namespace Beamable.Editor.Toolbox.Components
 
 			var installButton = Root.Q<Button>("announcement-install");
 			installButton.text = UpdateAvailableAnnouncementModel.InstallButtonText;
-			installButton.clickable.clicked += () => UpdateAvailableAnnouncementModel.OnInstall?.Invoke();
+			installButton.clickable.clicked += () =>
+			{
+				installButton.SetEnabled(false);
+				UpdateAvailableAnnouncementModel.OnInstall?.Invoke();
+			};
 		}
 	}
 }